Infer RootState type from RootReducer
Use ReturnType (introduced in TypeScript 2.8) to infer RootState's
type directly from the map used for constructing the RootReducer.

This prevents us from having to repeat ourselves, so when we add a
new reducer, we only need to add it in one place, so we can't
accidentally add it to the RootState type but forget to update the
reducer itself or vice-versa.

Upgrading the playground project to TS 2.8 caused issues with two
of the HOC examples because of their use of the spread operator,
which I was only able to resolve by using Object.assign in one
case, and using `as any` in another. Hopefully the TypeScript team
will fix this issue soon by introducing the `Spread` type.